LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-preload-and-consume.https-expected.txt


FAIL Reporting endpoints received credentials. assert_equals: expected 2 but got 1
